+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.3.1] - 2023-11-27
 ### Fixed
 - Allow responses without a content type (for example redirects).
 
@@ -50,7 +51,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - First implementation
 
-[Unreleased]: https://github.com/rcmachado/changelog/compare/v0.3.0...HEAD
+[Unreleased]: https://github.com/rcmachado/changelog/compare/v0.3.1...HEAD
+[0.3.1]: https://github.com/rcmachado/changelog/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/rcmachado/changelog/compare/v0.2.1...v0.3.0
 [0.2.1]: https://github.com/rcmachado/changelog/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/rcmachado/changelog/compare/v0.1.0...v0.2.0